TaskManager/pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["dbp"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
django-filter = "^23.1"
markdown = "^3.4.3"
drf-yasg = "^1.21.5"
djangorestframework-simplejwt = "^5.2.2"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
freezegun = "^1.2.2"
coverage = "^7.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"